requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==4.2.11
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-session-security==2.6.7
mysql-connector-python==8.4.0
mysqlclient==2.2.4
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
